Reorganized files and removed unnecessary imports
