+++ Entities/Map1.js
@@ -1,6 +1,6 @@
 import { GameMap } from "./Entities.js";
 import { Player } from "./Player/Player.js";
-import { Background } from "./Background.js";
+import { Background } from "./Map/Background.js";
 import { Cactus, Spider } from "./Enemy.js";
 import { Structure } from "./Structure.js";
 import { GAME_ENGINE } from "../main.js";
